move compare_on_tabarena logic to separate file
